Change railtie process
